requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
pandas>=1.5.0
Pillow>=9.0.0
numpy>=1.23.0
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
ipywidgets>=8.0.0
ipython>=8.0.0
